added circle and line on the board
